Create Cable-Tester.cpp
Created this awesome cable tester for Arduino/Atmega328 Microcontroller 
It works by outputting a known signal directly to 2 different ADC pins, one of which travels through the cable you are testing. 
The amplitude is measured between the signals and depending on the difference between the two, outputs a signal to an LED.

Still need to work on commenting out some logic and also eventually add a signal frequency sweep.