Create Multi-Dimensional-Tremolo.cpp
I've been working on this one for a while. There are a few bugs but they are unpredictable and some are perhaps a result of bad hardware setup. I haven't had the time/resources to do a hardware rebuild yet but when I do I want to start testing out this code again. One day it's time will come!